# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Duty scheduler API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Member {
  id: number;
  name: string;
  /** @nullable */
  color?: string | null;
  createdAt: string;
}

export interface MemberInput {
  /** @minLength 1 */
  name: string;
  color?: string;
}

export interface Location {
  id: number;
  name: string;
  createdAt: string;
}

export interface LocationInput {
  /** @minLength 1 */
  name: string;
}

export interface Schedule {
  id: number;
  title: string;
  /** @nullable */
  weekLabel?: string | null;
  createdAt: string;
}

export interface ScheduleInput {
  /** @minLength 1 */
  title: string;
  weekLabel?: string;
}

export interface AssignmentWithDetails {
  id: number;
  scheduleId: number;
  memberId: number;
  locationId: number;
  dayOfWeek: number;
  memberName: string;
  /** @nullable */
  memberColor?: string | null;
  locationName: string;
}

export interface ScheduleDetail {
  id: number;
  title: string;
  /** @nullable */
  weekLabel?: string | null;
  createdAt?: string;
  assignments: AssignmentWithDetails[];
}

export interface Assignment {
  id: number;
  scheduleId: number;
  memberId: number;
  locationId: number;
  /** 1=Monday, 2=Tuesday, 3=Wednesday, 4=Thursday, 5=Friday */
  dayOfWeek: number;
}

export interface AssignmentInput {
  memberId: number;
  locationId: number;
  dayOfWeek: number;
}