# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Duty scheduler API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: members
    description: Staff member operations
  - name: locations
    description: Location operations
  - name: schedules
    description: Schedule sheet operations
  - name: assignments
    description: Assignment operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /members:
    get:
      operationId: listMembers
      tags: [members]
      summary: List all staff members
      responses:
        "200":
          description: List of members
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Member"
    post:
      operationId: createMember
      tags: [members]
      summary: Create a new staff member
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MemberInput"
      responses:
        "201":
          description: Created member
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Member"

  /members/{id}:
    delete:
      operationId: deleteMember
      tags: [members]
      summary: Delete a staff member
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  /locations:
    get:
      operationId: listLocations
      tags: [locations]
      summary: List all locations
      responses:
        "200":
          description: List of locations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Location"
    post:
      operationId: createLocation
      tags: [locations]
      summary: Create a new location
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LocationInput"
      responses:
        "201":
          description: Created location
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Location"

  /locations/{id}:
    delete:
      operationId: deleteLocation
      tags: [locations]
      summary: Delete a location
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  /schedules:
    get:
      operationId: listSchedules
      tags: [schedules]
      summary: List all schedule sheets
      responses:
        "200":
          description: List of schedules
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Schedule"
    post:
      operationId: createSchedule
      tags: [schedules]
      summary: Create a new schedule sheet
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ScheduleInput"
      responses:
        "201":
          description: Created schedule
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Schedule"

  /schedules/{id}:
    get:
      operationId: getSchedule
      tags: [schedules]
      summary: Get a schedule sheet with all assignments
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Schedule with assignments
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ScheduleDetail"
        "404":
          description: Not found
    delete:
      operationId: deleteSchedule
      tags: [schedules]
      summary: Delete a schedule sheet
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  /schedules/{scheduleId}/assignments:
    post:
      operationId: createAssignment
      tags: [assignments]
      summary: Assign a member to a location/day slot
      parameters:
        - name: scheduleId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AssignmentInput"
      responses:
        "201":
          description: Created assignment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Assignment"

  /schedules/{scheduleId}/assignments/{id}:
    delete:
      operationId: deleteAssignment
      tags: [assignments]
      summary: Remove a member from a slot
      parameters:
        - name: scheduleId
          in: path
          required: true
          schema:
            type: integer
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    Member:
      type: object
      required: [id, name, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        color:
          type: ["string", "null"]
        createdAt:
          type: string

    MemberInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
          minLength: 1
        color:
          type: string

    Location:
      type: object
      required: [id, name, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        createdAt:
          type: string

    LocationInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
          minLength: 1

    Schedule:
      type: object
      required: [id, title, createdAt]
      properties:
        id:
          type: integer
        title:
          type: string
        weekLabel:
          type: ["string", "null"]
        createdAt:
          type: string

    ScheduleInput:
      type: object
      required: [title]
      properties:
        title:
          type: string
          minLength: 1
        weekLabel:
          type: string

    ScheduleDetail:
      type: object
      required: [id, title, assignments]
      properties:
        id:
          type: integer
        title:
          type: string
        weekLabel:
          type: ["string", "null"]
        createdAt:
          type: string
        assignments:
          type: array
          items:
            $ref: "#/components/schemas/AssignmentWithDetails"

    Assignment:
      type: object
      required: [id, scheduleId, memberId, locationId, dayOfWeek]
      properties:
        id:
          type: integer
        scheduleId:
          type: integer
        memberId:
          type: integer
        locationId:
          type: integer
        dayOfWeek:
          type: integer
          description: "1=Monday, 2=Tuesday, 3=Wednesday, 4=Thursday, 5=Friday"

    AssignmentInput:
      type: object
      required: [memberId, locationId, dayOfWeek]
      properties:
        memberId:
          type: integer
        locationId:
          type: integer
        dayOfWeek:
          type: integer

    AssignmentWithDetails:
      type: object
      required: [id, scheduleId, memberId, locationId, dayOfWeek, memberName, locationName]
      properties:
        id:
          type: integer
        scheduleId:
          type: integer
        memberId:
          type: integer
        locationId:
          type: integer
        dayOfWeek:
          type: integer
        memberName:
          type: string
        memberColor:
          type: ["string", "null"]
        locationName:
          type: string