# lib/api-zod/src/generated/types/assignment.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Duty scheduler API
 * OpenAPI spec version: 0.1.0
 */

export interface Assignment {
  id: number;
  scheduleId: number;
  memberId: number;
  locationId: number;
  /** 1=Monday, 2=Tuesday, 3=Wednesday, 4=Thursday, 5=Friday */
  dayOfWeek: number;
}