# lib/api-zod/src/generated/types/location.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Duty scheduler API
 * OpenAPI spec version: 0.1.0
 */

export interface Location {
  id: number;
  name: string;
  createdAt: string;
}