# lib/api-zod/src/generated/types/scheduleInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Duty scheduler API
 * OpenAPI spec version: 0.1.0
 */

export interface ScheduleInput {
  /** @minLength 1 */
  title: string;
  weekLabel?: string;
}